Guia 5: Calculador impositivo. Version testeada

+++ Guia 5/Calculador impositivo/include/Dinero.hpp
@@ -1,6 +1,6 @@
 #ifndef MONEY_HPP
 #define MONEY_HPP
 
-using Dinero = int;
+using Dinero = double;
 
 #endif
